+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # LessCSS Maven Plugin Changelog
 
+## 2.0.0
+Features:
+* Set compatibility with Java 6+
+* Set compatibility with Maven 3+
+
+Dependencies:
+* Upgraded [LessCSS Compiler](http://lesscss-compiler.projects.gabrys.biz/) from [1.2.2](http://lesscss-compiler.projects.gabrys.biz/1.2.2/) to [2.0.0](http://lesscss-compiler.projects.gabrys.biz/2.0.0/) ([#7](https://github.com/gabrysbiz/lesscss-maven-plugin/issues/7))
+* Removed [Extended LessCSS Compiler](http://lesscss-extended-compiler.projects.gabrys.biz/) ([#7](https://github.com/gabrysbiz/lesscss-maven-plugin/issues/7))
+* Upgraded [Maven Plugin Utils](http://maven-plugin-utils.projects.gabrys.biz/) from [1.4.0](http://maven-plugin-utils.projects.gabrys.biz/1.4.1/) to [2.0.0](http://maven-plugin-utils.projects.gabrys.biz/2.0.0/)
+
+[See documentation](http://lesscss-maven-plugin.projects.gabrys.biz/2.0.0/)
+
 ## 1.2.1
 Dependencies:
 * Use [LessCSS Compiler 1.2.2](http://lesscss-compiler.projects.gabrys.biz/1.2.2/) (fixed support for `data-uri` function - only for `local` compiler type)
@@ -8,13 +20,13 @@ Dependencies:
 
 ## 1.2.0
 Dependencies:
-* Upgrade [Extended LessCSS Compiler](http://lesscss-extended-compiler.projects.gabrys.biz/) from [2.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.0/) to [2.1.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.1.0/) (support for `classpath://` protocol)
+* Upgraded [Extended LessCSS Compiler](http://lesscss-extended-compiler.projects.gabrys.biz/) from [2.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.0/) to [2.1.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.1.0/) (support for `classpath://` protocol)
 
 [See documentation](http://lesscss-maven-plugin.projects.gabrys.biz/1.2.0/)
 
 ## 1.1
 Dependencies:
-* Upgrade [Extended LessCSS Compiler](http://lesscss-extended-compiler.projects.gabrys.biz/) from [1.0](http://lesscss-extended-compiler.projects.gabrys.biz/1.0/) to [2.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.0/)
+* Upgraded [Extended LessCSS Compiler](http://lesscss-extended-compiler.projects.gabrys.biz/) from [1.0](http://lesscss-extended-compiler.projects.gabrys.biz/1.0/) to [2.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.0/)
 
 [See documentation](http://lesscss-maven-plugin.projects.gabrys.biz/1.1/)